requirements.txt
Flask
firebase-admin
python-dotenv
cryptography
twilio
gunicorn